MakieMaestro

MakieMaestro attempts to add features for Makie to

Simplify theming consistency. This includes using different themes across different back-ends.

Make it easy to save-figures in different formats using different themes and with various back-ends to a given output directory. This also includes saving figures in all selected formats with one command.

Enable saving figures for a given physical size to be included in a document.

Simplifies overriding the theme for a specific case (e.g. margin figures, offset axes etc.) using override themes

and functions that generate them

Warning

Most of the features of this package is opinionated including the theming. It is made with the priority of optimizing my workflow.

See also: MakieExtra.jl

MakieMaestro.Themes

MakieMaestro.L

MakieMaestro.Recipes.image

MakieMaestro.Recipes.image

MakieMaestro.Recipes.image!

MakieMaestro.Recipes.image!

MakieMaestro.Recipes.mosaic

MakieMaestro.Themes.figsize

MakieMaestro.Themes.get_hwratio

MakieMaestro.Themes.get_theme

MakieMaestro.Themes.get_width

MakieMaestro.Themes.hwratio!

MakieMaestro.Themes.interactive_size!

MakieMaestro.Themes.merge_generate

MakieMaestro.Themes.screen_parameters

MakieMaestro.Themes.to_units

MakieMaestro.Themes.update_theme

MakieMaestro.Themes.update_theme!

MakieMaestro.Themes.width!

MakieMaestro.backend_formats

MakieMaestro.choose_backend

MakieMaestro.choose_backend

MakieMaestro.extension

MakieMaestro.extension

MakieMaestro.fftvis

MakieMaestro.figure_dir!

MakieMaestro.figure_dir!

MakieMaestro.get_figure_dir

MakieMaestro.get_figure_dir

MakieMaestro.get_themes

MakieMaestro.get_themes

MakieMaestro.isvectorgraphic

MakieMaestro.isvectorgraphic

MakieMaestro.savefig

MakieMaestro.savefig

MakieMaestro.skip

MakieMaestro.skip

MakieMaestro.L — Method

L(s::String)

Shortcut for latexstring(raw" ext{" * s * "}").

This is useful when you want to have LaTeX axis labels, but do not want to use the theme_latexfonts() theme.

MakieMaestro.Recipes.image! — Method

Recipes.image!(args...; decorations=false, interpolate=false, axis=(; aspect=DataAspect()), colormap=:viridis, kwargs...)

Modified image! recipe that enables the user to set axis attributes even with the mutating function and has different defaults.

decorations – hide the axis decorations if set to false

MakieMaestro.Recipes.image — Method

Recipes.image(f)

Helper for setting the prefered options (interpolate, ax.aspect, decorations) for the image recipe.

MakieMaestro.backend_formats — Method

MakieMaestro.backend_formats(backend::Module,...)

Return the set of compatible formats for given backends.

Examples

julia> @assert MakieMaestro.backend_formats(CairoMakie) == Set([ MakieMaestro.Png, MakieMaestro.Eps, MakieMaestro.Svg, MakieMaestro.Pdf ])

julia> @assert MakieMaestro.backend_formats(GLMakie) == Set([MakieMaestro.Png])

MakieMaestro.choose_backend — Method

MakieMaestro.choose_backend(backends::Vector{Module}, f::Format)

Select the best backend from backends to export a figure in the format f.

Prefers CairoMakie for most formats unless the format is Png and GLMakie is available.

Examples

julia> MakieMaestro.choose_backend([CairoMakie, GLMakie], MakieMaestro.Svg)
CairoMakie

julia> MakieMaestro.choose_backend([WGLMakie], MakieMaestro.Eps)
ERROR: None of the backends support the format Eps
[...]

MakieMaestro.extension — Method

MakieMaestro.extension(format::Format)

Return the string of the extension from the given format.

Examples

julia> MakieMaestro.extension(MakieMaestro.Png)
".png"

julia> MakieMaestro.extension(MakieMaestro.PdfTex)
".pdf"

MakieMaestro.fftvis — Method

fftvis(fft_img)

Apply a logarithmic scaling to visualize the absolute value of the Fourier image.

Extended help

This function takes the absolute value of the input img (typically the result of an FFT), adds 1 to avoid log(0), and then applies a base-2 logarithm. This scaling helps to visualize the wide dynamic range typically present in Fourier transforms of images.

Example

using FFTW
img = rand(100, 100)
fft_img = fft(img)
heatmap(fftvis(fftshift(fft_img)))

MakieMaestro.figure_dir! — Method

figure_dir!(dir)

Set the figure directory

Examples

julia> figure_dir!(".");

MakieMaestro.get_figure_dir — Method

MakieMaestro.get_figure_dir()

Get the figure directory

Examples

julia> figure_dir!("/tmp")
"/tmp"

julia> MakieMaestro.get_figure_dir()
"/tmp"

MakieMaestro.get_themes — Method

MakieMaestro.get_themes(backend::Module, format::Format)

Get the modification themes associated with the backend and format.

Examples

julia> MakieMaestro.get_themes(CairoMakie, MakieMaestro.PdfTex)
3-element Vector{Symbol}:
 :base
 :cairomakie
 :vector

julia> MakieMaestro.get_themes(GLMakie, MakieMaestro.Png)
3-element Vector{Symbol}:
 :base
 :glmakie
 :raster

MakieMaestro.isvectorgraphic — Method

MakieMaestro.isvectorgraphic(x::Format)

Return true if x is a vector graphic format, false otherwise.

Examples

julia> MakieMaestro.isvectorgraphic(MakieMaestro.Pdf)
true

julia> MakieMaestro.isvectorgraphic(MakieMaestro.Png)
false

MakieMaestro.savefig — Function

savefig(fig_function, name, dir; <keyword arguments>)

Save a figure output by fig_function in with themes applied and various formats in dir with the file name name

Parameters:

fig_function: function that generates the figure or figures to save

name: name of the file to save the figure as or vector of names if multiple figures are returned

dir: relative or absolute path to the project directory (default: FIGURE_DIR)

Save a figure in the selected formats. If :pdf_tex format is requested, Inkscape is used to convert the SVG file to PDF with text in LaTeX.

Keyword arguments:

hwratio=HWRATIO_DEFAULT

width=WIDTH_DEFAULT

backend=CairoMakie

override_theme=Theme()

size_theme=SIZE_THEME

base_theme=BASE_THEME

vector_theme=VECTOR_THEME

raster_theme=RASTER_THEME

gl_theme=GL_THEME

cairo_theme=CAIRO_THEME

skip=[:eps, :pdf_tex, :svg, :raster] - other options are :svg, :pdf, :pdf_tex, :eps, :png, :raster, :vector

fig_function_args=()

update=false

MakieMaestro.skip — Method

MakieMaestro.skip(skips::Vararg{Union{Symbol,Format}})

Return a Set{Format} of allowed formats by excluding specified formats or format groups.

Useful for customizing the output formats when saving figures, allowing you to easily exclude certain format types or groups of formats.

Arguments

skips: Variable number of arguments specifying formats or format groups to exclude. Can be individual Format types or symbols :raster or :vector.

Examples

julia> @assert MakieMaestro.skip(:raster) == Set([MakieMaestro.PdfTex, MakieMaestro.Eps, MakieMaestro.Svg, MakieMaestro.Pdf])

julia> @assert MakieMaestro.skip(:vector) == Set([MakieMaestro.Png])

julia> @assert MakieMaestro.skip(MakieMaestro.Pdf, MakieMaestro.Svg) == Set([MakieMaestro.PdfTex, MakieMaestro.Png, MakieMaestro.Eps])