MakieMaestro

MakieMaestro attempts to add features for Makie to

Simplify theming consistency. This includes using different themes across different back-ends.

Make it easy to save-figures in different formats using different themes and with various back-ends to a given output directory. This also includes saving figures in all selected formats with one command.

Enable saving figures for a given physical size to be included in a document.

Simplifies overriding the theme for a specific case (e.g. margin figures, offset axes etc.) using override themes

and functions that generate them

Warning

Most of the features of this package is opinionated including the theming. It is made with the priority of optimizing my workflow.

See also: MakieExtra.jl

MakieMaestro.L

MakieMaestro.Recipes.image

MakieMaestro.Recipes.image!

MakieMaestro.Themes.figsize

MakieMaestro.Themes.get_hwratio

MakieMaestro.Themes.get_theme

MakieMaestro.Themes.get_width

MakieMaestro.Themes.hwratio!

MakieMaestro.Themes.interactive_size!

MakieMaestro.Themes.merge_generate

MakieMaestro.Themes.screen_parameters

MakieMaestro.Themes.to_units

MakieMaestro.Themes.update_theme

MakieMaestro.Themes.update_theme!

MakieMaestro.Themes.width!

MakieMaestro.fftvis

MakieMaestro.figure_dir!

MakieMaestro.get_figure_dir

MakieMaestro.savefig

MakieMaestro.skip

MakieMaestro.vectorgraphic

MakieMaestro.L — Method

L(s::String)

Shortcut for latexstring(raw" ext{" * s * "}").

This is useful when you want to have LaTeX axis labels, but do not want to use the theme_latexfonts() theme.

MakieMaestro.Recipes.image! — Method

Recipes.image!(args...; decorations=false, interpolate=false, axis=(; aspect=DataAspect()), colormap=:viridis, kwargs...)

Modified image! recipe that enables the user to set axis attributes even with the mutating function and has different defaults.

decorations – hide the axis decorations if set to false

MakieMaestro.Recipes.image — Method

Recipes.image(f)

Helper for setting the prefered options (interpolate, ax.aspect, decorations) for the image recipe.

MakieMaestro.fftvis — Method

fftvis(fft_img)

Apply a logarithmic scaling to visualize the absolute value of the Fourier image.

Extended help

This function takes the absolute value of the input img (typically the result of an FFT), adds 1 to avoid log(0), and then applies a base-2 logarithm. This scaling helps to visualize the wide dynamic range typically present in Fourier transforms of images.

Example

using FFTW
img = rand(100, 100)
fft_img = fft(img)
heatmap(fftvis(fftshift(fft_img)))

MakieMaestro.figure_dir! — Method

figure_dir!(dir)

Set the figure directory

MakieMaestro.get_figure_dir — Method

get_figure_dir()

Get the figure directory

MakieMaestro.savefig — Function

savefig(fig_function, name, dir; <keyword arguments>)

Save a figure output by fig_function in with themes applied and various formats in dir with the file name name

Parameters:

fig_function: function that generates the figure or figures to save

name: name of the file to save the figure as or vector of names if multiple figures are returned

dir: relative or absolute path to the project directory (default: FIGURE_DIR)

Save a figure in the selected formats. If :pdf_tex format is requested, Inkscape is used to convert the SVG file to PDF with text in LaTeX.

Keyword arguments:

hwratio=HWRATIO_DEFAULT

width=WIDTH_DEFAULT

backend=CairoMakie

override_theme=Theme()

size_theme=SIZE_THEME

base_theme=BASE_THEME

vector_theme=VECTOR_THEME

raster_theme=RASTER_THEME

gl_theme=GL_THEME

cairo_theme=CAIRO_THEME

skip=[:eps, :pdf_tex, :svg, :raster] - other options are :svg, :pdf, :pdf_tex, :eps, :png, :raster, :vector

fig_function_args=()

update=false

MakieMaestro.skip — Method

skip(skips::Vararg{Union{Symbol,Format}})

Generate a set of allowed formats by excluding specified formats or format groups.

Arguments

skips: Variable number of arguments specifying formats or format groups to exclude. Can be individual Format types or symbols :raster or :vector.

Returns

A Set of allowed Format types after excluding the specified formats.

Examples

skip(:raster)  # Excludes Png format
skip(:vector)  # Excludes Png, Eps, PdfTex, and Svg formats
skip(Png, Svg) # Excludes Png and Svg formats specifically

This function is useful for customizing the output formats when saving figures, allowing you to easily exclude certain format types or groups of formats.

MakieMaestro.vectorgraphic — Method

vectorgraphic(x)

Determine if the given format is a vector graphic format.

Arguments

x: The format to check.

Returns

true if the format is a vector graphic format (Svg, Png, Eps, or PdfTex), false otherwise.

Examples

vectorgraphic(Svg)   # returns true
vectorgraphic(Jpg)   # returns false