Public Documentation

Documentation for MakieMaestro.jl's public interface.

Contents

Contents

Public Interface

RecipeOverrides

MakieMaestro.Recipes.image

MakieMaestro.Recipes.image!

MakieMaestro.Recipes.mosaic

MakieMaestro.Themes.figsize

MakieMaestro.Themes.get_hwratio

MakieMaestro.Themes.get_theme

MakieMaestro.Themes.get_width

MakieMaestro.Themes.hwratio!

MakieMaestro.Themes.interactive_size!

MakieMaestro.Themes.merge_generate

MakieMaestro.Themes.screen_parameters

MakieMaestro.Themes.to_units

MakieMaestro.Themes.update_theme

MakieMaestro.Themes.update_theme!

MakieMaestro.Themes.width!

Public Interface

MakieMaestro.Themes.get_theme — Function

get_theme(themes::Vector{Union{ThemeGenerator,Symbol}}; dict = THEME[])
get_theme(A, B, C, ...; dict = THEME[])

Create theme for specified keys and/or generators (including Themes).

Arguments

keys::Vector{Symbol}: A vector of symbols representing the desired theme property keys.

Example

gap = true
theme_props = get_theme([Theme(; figure_padding=2), :font, () -> Theme(; colgap = gap, rowgap = gap),  :linewidth, :color])

This will return a theme that combines the three specified themes together. Note that rightmost has precedence unlike merges on usual julia dictionaries, but same as with Makie.Attributes.

MakieMaestro.Themes.interactive_size! — Function

interactive_size!(size::Tuple{Length,Length})
interactive_size!(ratio::Tuple{Number,Number}; index=nothing)

Set the interactive figure size.

If given a ratio, calculate the size from the screen parameters so that the figure takes up the given width/height ratio of the screen. The optional index specifies the screen to use.

MakieMaestro.Themes.hwratio! — Function

hwratio!(val)

Set the height-width ratio for saving figures.

julia> hwratio!((1 + √(5))/2)
1.618033988749895

julia> Themes.get_hwratio()
1.618033988749895

MakieMaestro.Themes.get_hwratio — Function

Themes.get_hwratio()

Get the set height-width ratio for saving figures.

Examples

julia> hwratio!(0.5)
0.5

julia> Themes.get_hwratio()
0.5

MakieMaestro.Themes.get_width — Function

Themes.get_width()

Get the default figure width

Examples

julia> width!(10u"cm")
10 cm

julia> Themes.get_width()
10 cm

MakieMaestro.Themes.to_units — Function

Themes.to_units(val::Length)

Convert val to Makie figure units

Example

julia> Themes.to_units(2u"cm")
7200//127

MakieMaestro.Themes.figsize — Function

figsize(width::Length=get_width(), hw_ratio=get_hwratio())

Calculate the figure size in points (1u"pt" == (1//72)u"inch") based on the given width and height-to-width ratio.

Example

julia> w,h = Themes.figsize(32u"cm", (1 + √(5))/2)
(907, 1467)

MakieMaestro.Themes.screen_parameters — Function

screen_parameters() -> Vector{ScreenInfo}

Retrieve information about available screens using the xdpyinfo command.

This function parses the output of xdpyinfo to extract details about each screen, including its index, dimensions in pixels and millimeters, and whether it's the default screen.

Returns:

A vector of ScreenInfo objects, each containing details about a screen.

Note:

This function relies on the xdpyinfo command and is therefore only compatible with systems where this command is available (typically Unix-like systems with X11).

The function may return an empty vector if no screens are detected or if parsing fails.

Requires the Unitful.jl package for handling millimeter units.

Example:

screens = screen_parameters()
for screen in screens
    println("Screen $(screen.index): $(screen.size_px) pixels, $(screen.size_mm) physical size")
end

MakieMaestro.Themes.update_theme — Function

update_theme(key::Symbol, with::ThemeGenerator)

Update a specific theme component identified by key in the global THEME dictionary.

Arguments

key::Symbol: The key identifying the theme component to update.

with::ThemeGenerator: The new theme or function to update the existing theme with.

See also Use update_theme! for direct modifications.

Extended help

Specific behaviour for argument types:

If the key doesn't exist in THEME, it adds the new theme or function.

If the key exists:

For an existing Theme:

If with is a Function, it merges the result of with with the current theme.

If with is a Theme, it merges the current theme with with.

For an existing Function:

If with is a Theme, it creates a new function that merges with with the result of the current function.

If with is a Function, it throws an ArgumentError.

Throws

ArgumentError: If attempting to update a generating function with another generating

function.

MakieMaestro.Themes.update_theme! — Function

update_theme!(key::Symbol, new::ThemeGenerator)

Update a specific theme component in the global theme.

Arguments

key::Symbol: The key representing the theme component to be updated.

new::ThemeGenerator: The new theme generator to replace the existing one.

This function modifies the global theme by replacing the theme generator for the specified component with a new one. It directly updates the THEME global variable.

See also update_theme

MakieMaestro.Themes.width! — Function

width!(val::Length)

Set the default width for figures.

julia> width!(177u"mm" * 0.8)
141.6 mm

julia> Themes.get_width()
141.6 mm

MakieMaestro.Themes.merge_generate — Function

merge_generate(theme::ThemeGenerator,...)::Theme

Generate and merge all the theme arguments.

Later input arguments have precedence similarly as in a merge of dictionaries. ``See also [get_theme](@ref), [ThemeGenerator`](@ref)

RecipeOverrides

MakieMaestro.Recipes.image! — Function

Recipes.image!(args...; decorations=false, interpolate=false, axis=(; aspect=DataAspect()), colormap=:viridis, kwargs...)

Modified image! recipe that enables the user to set axis attributes even with the mutating function and has different defaults.

decorations – hide the axis decorations if set to false

MakieMaestro.Recipes.image — Function

Recipes.image(f)

Helper for setting the prefered options (interpolate, ax.aspect, decorations) for the image recipe.

Helper plotting functions

MakieMaestro.Recipes.mosaic — Function

mosaic(datas::AbstractVector; 
        nrows, ncols,
       plt=Recipes.image!, ax_func! = identity, linkaxes=true, vargs...)
mosaic(datas::Stack, args...; vargs...)
mosaic(datas::AbstractMatrix; vargs...)

Create a grid-like arrangement of plots.

Arguments

datas: Vector, 3D array, or matrix of data to be plotted

nrows / ncols: Number of rows / columns

plt: Plotting function to apply (default: image!)

ax!: Function to modify axis properties (default: identity)

linkaxes: Boolean to control axis linking (default: true)

vargs...: Additional plotting arguments passed to the plotting function

The function creates a Makie Figure object with a grid of plots, applies the specified plotting function to each data element, and returns the resulting figure. It supports vectorized arguments that match the data length and automatically handles tuple-packed data.

Returns

A Makie Figure object containing the mosaic of plots that have linked axes

Examples

# Basic usage with vector of matrices
data = [rand(10,10) for i in 1:4]
fig = mosaic(data, 2, 2)

# Using with a 3D array
data3d = rand(10,10,4) 
fig = mosaic(data3d)

# Using with a matrix
datamat = [rand(2,3) for _ in 1:3, _ in 1:4]
fig = mosaic(datamat)