Public Documentation

Documentation for MakieMaestro.jl's public interface.

Contents

Contents

Logging

Paper Sizes

Overrides

Utility / QOL functions

Exporting

MakieMaestro.savefig — Function

savefig(fig, [path], [size]; <keyword arguments>)

Save the figure fig at path with size.

Themes are applied from the :base theme and you can supply any number of overrides, either predefined or you own with the override_theme keyword.

The arguments fig, path and size may be specified in a number of different ways as shown below in Arguments or in the case of path and size be left with their default values (see Themes.width!, Themes.hwratio!, figure_dir! and export_format!). For a clearer picture of how to supply the arguments, see Examples in the documentation, but as a rule of thumb, any sensible way to define the export arguments should work granted that they are in the correct order.

Arguments

fig – figure(s) generating function

Info

fig_func::Function, (arg1, arg2,...), () -> fig_func(arg1, arg2,...) or fig_func (if it is possible to call it without any arguments). This is the only argument that is necessary to provide. See all possible signatures at FunctionSpec.

path: basename, directory and formats

Info

name, [formats], [dir] or path. For example "protein_density_heatmap", [:svg, :pdf, :pdf_tex], "~/ImporantProject/" or equivalently "~/ImporantProject/protein_density_heatmap.{svg, pdf, pdf_tex}". See all possible signatures at PathSpec.

Tip

If not supplied, the name is inferred from nameof(fig_func), formats are taken from MakieMaestro.get_export_format (see also export_format!) and dir is taken from MakieMaestro.get_figure_dir (see also figure_dir!).

Warning

If PdfTex format is requested, Inkscape is used to convert the SVG file to PDF with text in LaTeX. Due to that incscape must be an executable on your system if you want to use the PdfTex format.

size: physical dimensions of the figure

Info

20.5u"cm", 0.6, FigHeight(5u"inch"), 8u"inch", 0.5, 1.5 (relative width, hwratio), HeightLength(0.3), 0.4 (relative height, hwratio). A combination of width, height and/or hwratio, such that the size is completely determined from these and the default values (see Themes.hwratio! and Themes.width!). See all possible signatures at SizeSpec.

Keyword arguments

backends=CairoMakie – Backend used for exporting

override_theme=Theme() – Theme(s) that will be used to style the figure

[update] – (default: determine by the backend)

kwargs... – Any additional keyword arguments are supplied to the figure function as keyword arguments.

MakieMaestro.figure_dir! — Function

figure_dir!(dir)

Set the figure directory

Examples

julia> figure_dir!(".");

MakieMaestro.export_format! — Function

export_format!(formats)

Set the default formats to export.

julia> export_format!(["pdf", :eps]);

julia> export_format!("pdf", MakieMaestro.Png, :svg);

julia> export_format!(missing);

Internals

MakieMaestro.isvectorgraphic — Function

MakieMaestro.isvectorgraphic(x::Format)

Return true if x is a vector graphic format, false otherwise.

Examples

julia> MakieMaestro.isvectorgraphic(MakieMaestro.Pdf)
true

julia> MakieMaestro.isvectorgraphic(MakieMaestro.Png)
false

MakieMaestro.skip — Function

MakieMaestro.skip(skips::Vararg{Union{Symbol,Format}})

Return a Set{Format} of allowed formats by excluding specified formats or format groups.

Useful for customizing the output formats when saving figures, allowing you to easily exclude certain format types or groups of formats.

Arguments

skips: Variable number of arguments specifying formats or format groups to exclude. Can be individual Format types or symbols :raster or :vector.

Examples

julia> @assert MakieMaestro.skip(:raster) == Set([MakieMaestro.PdfTex, MakieMaestro.Eps, MakieMaestro.Svg, MakieMaestro.Pdf])

julia> @assert MakieMaestro.skip(:vector) == Set([MakieMaestro.Png])

julia> @assert MakieMaestro.skip(MakieMaestro.Pdf, MakieMaestro.Svg) == Set([MakieMaestro.PdfTex, MakieMaestro.Png, MakieMaestro.Eps])

MakieMaestro.extension — Function

MakieMaestro.extension(format::Format)

Return the string of the extension from the given format.

Examples

julia> MakieMaestro.extension(MakieMaestro.Png)
".png"

julia> MakieMaestro.extension(MakieMaestro.PdfTex)
".pdf"

MakieMaestro.choose_backend — Function

MakieMaestro.choose_backend(backends::Vector{Module}, f::Format)

Select the best backend from backends to export a figure in the format f.

Prefers CairoMakie for most formats unless the format is Png and GLMakie is available.

Examples

julia> MakieMaestro.choose_backend([CairoMakie, GLMakie], MakieMaestro.Svg)
CairoMakie

julia> MakieMaestro.choose_backend([WGLMakie], MakieMaestro.Eps)
ERROR: None of the backends support the format Eps
[...]

MakieMaestro.backend_formats — Function

MakieMaestro.backend_formats(backend::Module,...)

Return the set of compatible formats for given backends.

Examples

julia> @assert Set([ MakieMaestro.Png, MakieMaestro.Eps, MakieMaestro.Svg, MakieMaestro.Pdf ]) ⊆ MakieMaestro.backend_formats(CairoMakie) 

julia> @assert MakieMaestro.backend_formats(GLMakie) == Set([MakieMaestro.Png])

MakieMaestro.get_figure_dir — Function

MakieMaestro.get_figure_dir()

Get the figure directory

Examples

julia> figure_dir!("/tmp")
"/tmp"

julia> MakieMaestro.get_figure_dir()
"/tmp"

MakieMaestro.get_themes — Function

MakieMaestro.get_themes(backend::Module, format::Format)

Get the modification themes associated with the backend and format.

Examples

julia> MakieMaestro.get_themes(CairoMakie, MakieMaestro.PdfTex)
2-element Vector{Symbol}:
 :cairomakie
 :vector

julia> MakieMaestro.get_themes(GLMakie, MakieMaestro.Png)
2-element Vector{Symbol}:
 :glmakie
 :raster

MakieMaestro.get_export_theme — Function

MakieMaestro.get_export_theme(backend, format, width, hwratio, override_theme...)

Create the final export theme applied when generating the figure.

Examples

julia> theme = MakieMaestro.get_export_theme(CairoMakie, MakieMaestro.Png, 10u"cm", 0.8,
           Theme(;
               CairoMakie=(;pdf_version=1.1)
           ),
           Theme(;
               Scatter=(;markersize=8)
           )
       );

julia> theme.size[]
(283, 226)

MakieMaestro.get_formats — Function

MakieMaestro.get_formats(backends::Vector{Module}, formats::Set{Format})

Return the list (Vector) of formats to generate based on the given backends required formats. Output is in the correct order for export.

Examples

julia> (Sys.which("inkscape") === nothing) || MakieMaestro.Svg ∈ MakieMaestro.get_formats([CairoMakie], Set([MakieMaestro.PdfTex])) # Added Svg so the PdfTex is possible to create
true

julia> MakieMaestro.get_formats([GLMakie], Set([MakieMaestro.Pdf]))
ERROR: ArgumentError: None of the backends Module[GLMakie] support the formats Set(MakieMaestro.Format[MakieMaestro.Pdf])
[...]

julia> MakieMaestro.get_formats([CairoMakie, GLMakie], Set([MakieMaestro.Png, MakieMaestro.Svg]))
2-element Vector{MakieMaestro.Format}:
 Svg::Format = 1
 Png::Format = 0

MakieMaestro.get_export_format — Function

MakieMaestro.get_export_format()

Get the default export formats

Examples

julia> export_format!(:pdf)
Set{MakieMaestro.Format} with 1 element:
  MakieMaestro.Pdf

julia> MakieMaestro.get_export_format()
Set{MakieMaestro.Format} with 1 element:
  MakieMaestro.Pdf

julia> export_format!(missing);

MakieMaestro.FunctionSpec — Type

FunctionSpec(fig, args)

Specification of the function generating the figure with its arguments.

Can be called with additional keyword arguments. The arguments are passed from the args field.

MakieMaestro.uniqueid — Function

uniqueid(fspec::FunctionSpec; kwargs...)

Generate a unique ID for the fspec when called with kwargs.

Useful for caching the figures that do not change their definitions so that they do not have to be re-generated when re-exported.

See also uniqueids.

julia> f(args...; kwargs...) = lines(args...; kwargs...);

julia> fspec1 = FunctionSpec(f, (1:10, 1:10));

julia> fspec2 = FunctionSpec(f, (1:10, 1:10));

julia> fspec3 = FunctionSpec(f, (1:20, 1:20));

julia> @assert uniqueid(fspec1) == uniqueid(fspec2)

julia> @assert uniqueid(fspec1) != uniqueid(fspec1; axis=(;title="Line")) 

julia> @assert uniqueid(fspec1) != uniqueid(fspec3)

MakieMaestro.uniqueids — Function

uniqueids(fspec::FunctionSpec; kwargs...)

Returns a named tuple of the unique IDs for the fspec when called with kwargs.

See also uniqueid.

julia> f(args...; kwargs...) = lines(args...; kwargs...);

julia> fspec1 = FunctionSpec(f, (1:10, 1:10));

julia> fspec2 = FunctionSpec(f, (1:10, 1:10));

julia> @assert uniqueids(fspec1).code == uniqueids(fspec2).code

julia> @assert uniqueids(fspec1; axis=(;title="Line")).kwargs != uniqueids(fspec2).kwargs

julia> fspec3 = FunctionSpec(f, (1:20, 1:20));

julia> @assert uniqueids(fspec1).code == uniqueids(fspec3).code

julia> @assert uniqueids(fspec1).args != uniqueids(fspec3).args

MakieMaestro.PathSpec — Type

PathSpec(basename, [formats], [dirname])
PathSpec(fullpath, [formats])

Specification of where and which format to use for exporting a figure.

basename is internally a function that takes an index integer and returns the name of the figure. The function is created either by a suffix or by parsing the supplied string (see examples below).

formats is a set or collection of formats wanted for the export.

dirname is the directory where the figure will be saved (checked for existence upon creation).

Examples

```jldoctest; setup=:(using MakieMaestro: PathSpec) julia> PathSpec("somefile.pdf", ".") ./somefile.pdf

julia> PathSpec("somefile.{svg,eps}", "..") ../somefile.{eps,svg}

julia> PathSpec("somefile.png", "nonexistentdir") ERROR: ArgumentError: `nonexistent_dir` is not a valid directory [...]

julia> PathSpec("somefile",["pdf", :svg], "..") ../somefile.{pdf,svg}

MakieMaestro.SizeSpec — Type

MakieMaestro.SizeSpec(width::Length, hwratio::Real)

Specification of the physical size of the exported image.

The primary way to represent the size of a figure is it's width and the height to width ratio hwratio. However, any set of two values from {width, hwratio, height} fully determine the size. width and height may be either fixed values of the type Length or relative sizes to their default values by the type Real (RelativeSize). None, one or two of the values may be supplied and the size is determined from these values and the defaults. These defaults may be set with width! and hwratio!.

See also Themes.width!, Themes.get_width, Themes.hwratio!, Themes.get_hwratio

Examples

julia> Themes.width!(Themes.A4_WIDTH), Themes.hwratio!(2/(√(5) + 1));

julia> SizeSpec(0.5, 2)
w×h: 105.0 mm × 210.0 mm [h/w: 2.0]

julia> SizeSpec(10u"cm")
w×h: 10.0 cm × 6.18 cm [h/w: 2/(√5 + 1)]

julia> SizeSpec((15u"cm", 10u"cm"))
w×h: 10.0 cm × 15.0 cm [h/w: 1.5]

julia> SizeSpec((15u"cm", 0.5))
w×h: 105.0 mm × 150.0 mm [h/w: 1.4]

julia> SizeSpec((2, 10u"cm"))
w×h: 10.0 cm × 12.36 cm [h/w: 1.2]

MakieMaestro.RelativeSize — Type

MakieMaestro.RelativeSize(ratio::Real)

Size relative to the default or calculated dimensions.

MakieMaestro.FigHeight — Type

FigHeight(height::Length)
FigHeight(height::Real) # relative size to the default

Specifies the height of the exported figure.

See also RelativeSize.

MakieMaestro._savefig — Function

_savefig(fig, name, format, backend, dir; update)

Save figure fig with backend and name in dir with format.

Internal function that gets called at the end of the exporting stack with all of the arguments already fully determined and the figure created with the theme activated. update should be true if the Axis is created separately from the plots contained in the axis in order to set the correct viewing limits for the figure.

MakieMaestro._savepdftex — Function

_savepdftex(svgpath, outputpath; wait=true)

Run the command for creating a PDFTEX figure using Inkscape.

Internal function to convert SVG figures to PDFTEX (PDF+LaTeX) format assuming that the SVG already exists. If wait then the command in ran as blocking.

Logging

Internals

MakieMaestro.create_logger — Function

create_logger(logger, figdir)

Create a logger based on the logger argument and the figure directory figdir.

If logger is true, a default logger is created, if logger is an AbstractString, a TOMLLogger at that path. If logger is false the NullLogger is created.

MakieMaestro.FigureLogger — Type

FigureLogger

An abstract type that can log the figures that are created in a directory.

MakieMaestro.add_entry — Method

add_entry(logger::FigureLogger, key::String, data::AbstractDict; update=true, keep=true)

Add the entry data under the key in the log file.

MakieMaestro.TOMLLogger — Type

TOMLLogger

A logger that logs the figure data in a TOML file.

MakieMaestro.add_entry — Method

add_entry(logger::TOMLLogger, key::String, data::AbstractDict; update=true, merge=true)

Add the entry data under the key in the log file.

If update is true, the entry under key will be updated if it already exists. If merge is true, the data will be merged with the existing entry.

MakieMaestro.logs_path — Method

logs_path(l::TOMLLogger)

Get the path of the log file used by l

MakieMaestro.inlineids — Function

inlineids(d::AbstractDict)

Collect the InlineDict into an IdSet.

MakieMaestro.NullLogger — Type

NullLogger

A logger that does nothing.

MakieMaestro.InlineDict — Type

InlineDict{K,V}

A light wrapper around the parent::AbstractDict that tells the logger to log inline.

Theming

MakieMaestro.Themes — Module

This module is used to set themes for your figures. It can help you define a consistent theming scheme for your saved figures and for the figures you experiment with interactively.

It intends to solve the problem, where Makie gives you the option of theming, but if you want to change the theme at multiple places with multiple attribute changes and perhaps combine these options differently for your figures, it is necessary that you create the override themes and merge them with the current one which sooner or later becomes tedious and error prone. Especially if you are to create a lot of figures with various theming preferences. This tries to make theming and keeping consistency in your plots a little bit easier by allowing you to define all the connected attributes in one theme, register it with update_theme!(:my_theme, theme) and use these where necessary in your figure creation process and when saving.

MakieMaestro.Themes.width! — Function

width!(val::Length)

Set the default width for figures.

julia> width!(177u"mm" * 0.8)
141.6 mm

julia> Themes.get_width()
141.6 mm

MakieMaestro.Themes.hwratio! — Function

hwratio!(val)

Set the height-width ratio for saving figures.

julia> hwratio!((1 + √(5))/2)
1.618033988749895

julia> Themes.get_hwratio()
1.618033988749895

MakieMaestro.Themes.update_theme — Function

update_theme(key::Symbol, with::ThemeGenerator)

Update a specific theme component identified by key in the global THEME dictionary.

Arguments

key::Symbol: The key identifying the theme component to update.

with::ThemeGenerator: The new theme or function to update the existing theme with.

See also update_theme! for direct modifications.

Extended help

Specific behaviour for argument types:

If the key doesn't exist in THEME, it adds the new theme or function.

If the key exists:

For an existing Theme:

If with is a Function, it merges the result of with with the current theme.

If with is a Theme, it merges the current theme with with.

For an existing Function:

If with is a Theme, it creates a new function that merges with with the result of the current function.

If with is a Function, it throws an ArgumentError.

Throws

ArgumentError: If attempting to update a generating function with another generating

function.

MakieMaestro.Themes.update_theme! — Function

update_theme!(key::Symbol, new::ThemeGenerator)

Update a specific theme component in the global theme.

Arguments

key::Symbol: The key representing the theme component to be updated.

new::ThemeGenerator: The new theme generator to replace the existing one.

This function modifies the global theme by replacing the theme generator for the specified component with a new one. It directly updates the THEME global variable.

See also update_theme

MakieMaestro.Themes.get_theme — Function

get_theme(themes::Vector{Union{ThemeGenerator,Symbol}}; dict = THEME[])
get_theme(A, B, C, ...; dict = THEME[])

Create theme for specified keys and/or generators (including Themes).

Later arguments have precedence similarly to Base.merge(A::Dict, B::Dict).

Arguments

keys: A Vector or Vararg of symbols, functions generating themes or themes. Symbols are taken as keys from the dict.

Examples

julia> gap = true
true

julia> theme_props = MakieMaestro.Themes.get_theme([Theme(; figure_padding=2), () -> Theme(; colgap = gap, rowgap = gap), :orange_title]);

julia> @assert theme_props.colgap[] == theme_props.rowgap[] == true

MakieMaestro.Themes.theme_keys — Function

theme_keys()

Show the that are defined in the theme dictionary.

julia> theme_keys()[1:3]
3-element Vector{Symbol}:
 :base
 :cairomakie
 :data_aspect

Internals

MakieMaestro.Themes.ThemeGenerator — Type

ThemeGenerator == Union{Function,Theme}

A struct used for defining new theme “mixins” that can be inferred from some latent parameters.

Danger

That the function returns a Theme value is not checked. Therefore this type is only to be used internally...

julia> (() -> nothing) isa MakieMaestro.Themes.ThemeGenerator
true

MakieMaestro.Themes.gen — Function

gen(s::Symbol; dict=THEME[])

Return the function from the key s of dict.

MakieMaestro.Themes.merge_generate — Function

merge_generate(theme::ThemeGenerator,...)::Theme

Generate and merge all the theme arguments.

Later input arguments have precedence similarly as in a merge of dictionaries.

See also get_theme, ThemeGenerator

Examples

julia> MakieMaestro.Themes.merge_generate(Theme(; figure_padding=3), () -> Theme(; figure_padding=6))
Attributes with 1 entry:
  figure_padding => 6

MakieMaestro.Themes.get_hwratio — Function

Themes.get_hwratio()

Get the set height-width ratio for saving figures.

Examples

julia> hwratio!(0.5)
0.5

julia> Themes.get_hwratio()
0.5

MakieMaestro.Themes.get_width — Function

Themes.get_width()

Get the default figure width

Examples

julia> width!(10u"cm")
10 cm

julia> Themes.get_width()
10 cm

MakieMaestro.Themes.to_units — Function

Themes.to_units(val::Length)

Convert val to Makie figure units

Example

julia> Themes.to_units(2u"cm")
7200//127

MakieMaestro.Themes.figsize — Function

figsize(width::Length=get_width(), hw_ratio=get_hwratio())

Calculate the figure size in points (1u"pt" == (1//72)u"inch") based on the given width and height-to-width ratio.

Examples

julia> w,h = Themes.figsize(32u"cm", (1 + √(5))/2)
(907, 1467)

MakieMaestro.Themes.ScreenInfo — Type

MakieMaestro.Themes.ScreenInfo

Fields

default::Union{Missing,Bool} – X11 default screen

index::Int – X11 screen index

size::Tuple{Int,Int} – pixel dimensions of the display

dimensions::Tuple{Length,Length} – physical dimensions of the display

MakieMaestro.Themes.interactive_size! — Function

interactive_size!(size::Tuple{Length,Length})
interactive_size!(ratio::Tuple{Number,Number}; index=nothing)

Set the interactive figure size.

If given a ratio, calculate the size from the screen parameters so that the figure takes up the given width/height ratio of the screen. The optional index specifies the screen to use.

See also screen_parameters, get_interactive_size

MakieMaestro.Themes.get_interactive_size — Function

get_interactive_size()

Retrieve the current interactive size setting for figures.

This function checks the global INTERACTIVE_SIZE variable and returns its value. If INTERACTIVE_SIZE is not set (i.e., it's missing), it attempts to set a default size of 80% of the screen dimensions using interactive_size!((0.8, 0.8)).

See also interactive_size!

Throws

ErrorException: If INTERACTIVE_SIZE is not set and the function fails to set a default value.

MakieMaestro.Themes.screen_parameters — Function

screen_parameters()

Retrieve information about available screens (Vector{ScreenInfo}) using the xdpyinfo command.

This function parses the output of xdpyinfo to extract details about each screen, including its index, dimensions in pixels and millimeters, and whether it's the default screen.

Examples

screens = screen_parameters()
for screen in screens
    println("Screen $(screen.index): $(screen.size) pixels, $(screen.dimensions) physical size")
end
# Screen 0: (4080, 1920) pixels, (1072 mm, 504 mm) physical size

See also ScreenInfo

Extended help

This function relies on the xdpyinfo command and is therefore only compatible with systems where this command is available (typically Unix-like systems with X11).

The function may return an empty vector if no screens are detected or if parsing fails.

Requires the Unitful.jl package for handling millimeter units.

Paper Sizes

ISO 216 standard

The standard under ISO 216 is to define the paper size of {A/B/C}{1-X} paper to be the size when the paper with an index one lower is folded in half. This means that with an increasing index we alternate between the width > height and width < height, or, loosely speaking, "landscape" and "portrait" orientation. If you truly want to use the standard, then the constants {A/B/C}{1-X}_{WIDTH/HEIGHT}_TRUE are the ones that you should use. If like any other sane person, you view the width as the left to right span of the paper, when oriented vertically you should use the constants {A/B/C}{1-X}_{WIDTH/HEIGHT} instead.

using MakieMaestro.Themes
@assert Themes.A0_WIDTH == Themes.A1_HEIGHT
@assert Themes.A2_HEIGHT_TRUE == Themes.A1_WIDTH_TRUE # weird

Paper Size Constants

MakieMaestro.Themes.A0_WIDTH — Constant

const A0_WIDTH = 841.0 mm

Width of the A0 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A0_WIDTH_TRUE, A0_HEIGHT

MakieMaestro.Themes.A0_WIDTH_TRUE — Constant

const A0_WIDTH_TRUE = 841.0 mm

Width of the A0 paper under the ISO 216 standard.

See also A0_WIDTH_TRUE, A0_WIDTH

MakieMaestro.Themes.A0_HEIGHT — Constant

const A0_HEIGHT = 1189.0 mm

Height of the A0 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A0_HEIGHT_TRUE, A0_WIDTH

MakieMaestro.Themes.A0_HEIGHT_TRUE — Constant

const A0_HEIGHT_TRUE = 1189.0 mm

Height of the A0 paper under the ISO 216 standard.

See also A0_HEIGHT_TRUE, A0_HEIGHT

MakieMaestro.Themes.A1_WIDTH — Constant

const A1_WIDTH = 594.0 mm

Width of the A1 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A1_WIDTH_TRUE, A1_HEIGHT

MakieMaestro.Themes.A1_WIDTH_TRUE — Constant

const A1_WIDTH_TRUE = 594.5 mm

Width of the A1 paper under the ISO 216 standard.

See also A1_WIDTH_TRUE, A1_WIDTH

MakieMaestro.Themes.A1_HEIGHT — Constant

const A1_HEIGHT = 841.0 mm

Height of the A1 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A1_HEIGHT_TRUE, A1_WIDTH

MakieMaestro.Themes.A1_HEIGHT_TRUE — Constant

const A1_HEIGHT_TRUE = 841.0 mm

Height of the A1 paper under the ISO 216 standard.

See also A1_HEIGHT_TRUE, A1_HEIGHT

MakieMaestro.Themes.A2_WIDTH — Constant

const A2_WIDTH = 420.0 mm

Width of the A2 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A2_WIDTH_TRUE, A2_HEIGHT

MakieMaestro.Themes.A2_WIDTH_TRUE — Constant

const A2_WIDTH_TRUE = 420.5 mm

Width of the A2 paper under the ISO 216 standard.

See also A2_WIDTH_TRUE, A2_WIDTH

MakieMaestro.Themes.A2_HEIGHT — Constant

const A2_HEIGHT = 594.0 mm

Height of the A2 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A2_HEIGHT_TRUE, A2_WIDTH

MakieMaestro.Themes.A2_HEIGHT_TRUE — Constant

const A2_HEIGHT_TRUE = 594.5 mm

Height of the A2 paper under the ISO 216 standard.

See also A2_HEIGHT_TRUE, A2_HEIGHT

MakieMaestro.Themes.A3_WIDTH — Constant

const A3_WIDTH = 297.0 mm

Width of the A3 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A3_WIDTH_TRUE, A3_HEIGHT

MakieMaestro.Themes.A3_WIDTH_TRUE — Constant

const A3_WIDTH_TRUE = 297.25 mm

Width of the A3 paper under the ISO 216 standard.

See also A3_WIDTH_TRUE, A3_WIDTH

MakieMaestro.Themes.A3_HEIGHT — Constant

const A3_HEIGHT = 420.0 mm

Height of the A3 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A3_HEIGHT_TRUE, A3_WIDTH

MakieMaestro.Themes.A3_HEIGHT_TRUE — Constant

const A3_HEIGHT_TRUE = 420.5 mm

Height of the A3 paper under the ISO 216 standard.

See also A3_HEIGHT_TRUE, A3_HEIGHT

MakieMaestro.Themes.A4_WIDTH — Constant

const A4_WIDTH = 210.0 mm

Width of the A4 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A4_WIDTH_TRUE, A4_HEIGHT

MakieMaestro.Themes.A4_WIDTH_TRUE — Constant

const A4_WIDTH_TRUE = 210.25 mm

Width of the A4 paper under the ISO 216 standard.

See also A4_WIDTH_TRUE, A4_WIDTH

MakieMaestro.Themes.A4_HEIGHT — Constant

const A4_HEIGHT = 297.0 mm

Height of the A4 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A4_HEIGHT_TRUE, A4_WIDTH

MakieMaestro.Themes.A4_HEIGHT_TRUE — Constant

const A4_HEIGHT_TRUE = 297.25 mm

Height of the A4 paper under the ISO 216 standard.

See also A4_HEIGHT_TRUE, A4_HEIGHT

MakieMaestro.Themes.A5_WIDTH — Constant

const A5_WIDTH = 148.0 mm

Width of the A5 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A5_WIDTH_TRUE, A5_HEIGHT

MakieMaestro.Themes.A5_WIDTH_TRUE — Constant

const A5_WIDTH_TRUE = 148.625 mm

Width of the A5 paper under the ISO 216 standard.

See also A5_WIDTH_TRUE, A5_WIDTH

MakieMaestro.Themes.A5_HEIGHT — Constant

const A5_HEIGHT = 210.0 mm

Height of the A5 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A5_HEIGHT_TRUE, A5_WIDTH

MakieMaestro.Themes.A5_HEIGHT_TRUE — Constant

const A5_HEIGHT_TRUE = 210.25 mm

Height of the A5 paper under the ISO 216 standard.

See also A5_HEIGHT_TRUE, A5_HEIGHT

MakieMaestro.Themes.A6_WIDTH — Constant

const A6_WIDTH = 105.0 mm

Width of the A6 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A6_WIDTH_TRUE, A6_HEIGHT

MakieMaestro.Themes.A6_WIDTH_TRUE — Constant

const A6_WIDTH_TRUE = 105.125 mm

Width of the A6 paper under the ISO 216 standard.

See also A6_WIDTH_TRUE, A6_WIDTH

MakieMaestro.Themes.A6_HEIGHT — Constant

const A6_HEIGHT = 148.0 mm

Height of the A6 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A6_HEIGHT_TRUE, A6_WIDTH

MakieMaestro.Themes.A6_HEIGHT_TRUE — Constant

const A6_HEIGHT_TRUE = 148.625 mm

Height of the A6 paper under the ISO 216 standard.

See also A6_HEIGHT_TRUE, A6_HEIGHT

MakieMaestro.Themes.A7_WIDTH — Constant

const A7_WIDTH = 74.0 mm

Width of the A7 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A7_WIDTH_TRUE, A7_HEIGHT

MakieMaestro.Themes.A7_WIDTH_TRUE — Constant

const A7_WIDTH_TRUE = 74.3125 mm

Width of the A7 paper under the ISO 216 standard.

See also A7_WIDTH_TRUE, A7_WIDTH

MakieMaestro.Themes.A7_HEIGHT — Constant

const A7_HEIGHT = 105.0 mm

Height of the A7 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A7_HEIGHT_TRUE, A7_WIDTH

MakieMaestro.Themes.A7_HEIGHT_TRUE — Constant

const A7_HEIGHT_TRUE = 105.125 mm

Height of the A7 paper under the ISO 216 standard.

See also A7_HEIGHT_TRUE, A7_HEIGHT

MakieMaestro.Themes.A8_WIDTH — Constant

const A8_WIDTH = 52.0 mm

Width of the A8 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A8_WIDTH_TRUE, A8_HEIGHT

MakieMaestro.Themes.A8_WIDTH_TRUE — Constant

const A8_WIDTH_TRUE = 52.5625 mm

Width of the A8 paper under the ISO 216 standard.

See also A8_WIDTH_TRUE, A8_WIDTH

MakieMaestro.Themes.A8_HEIGHT — Constant

const A8_HEIGHT = 74.0 mm

Height of the A8 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A8_HEIGHT_TRUE, A8_WIDTH

MakieMaestro.Themes.A8_HEIGHT_TRUE — Constant

const A8_HEIGHT_TRUE = 74.3125 mm

Height of the A8 paper under the ISO 216 standard.

See also A8_HEIGHT_TRUE, A8_HEIGHT

MakieMaestro.Themes.A9_WIDTH — Constant

const A9_WIDTH = 37.0 mm

Width of the A9 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A9_WIDTH_TRUE, A9_HEIGHT

MakieMaestro.Themes.A9_WIDTH_TRUE — Constant

const A9_WIDTH_TRUE = 37.15625 mm

Width of the A9 paper under the ISO 216 standard.

See also A9_WIDTH_TRUE, A9_WIDTH

MakieMaestro.Themes.A9_HEIGHT — Constant

const A9_HEIGHT = 52.0 mm

Height of the A9 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A9_HEIGHT_TRUE, A9_WIDTH

MakieMaestro.Themes.A9_HEIGHT_TRUE — Constant

const A9_HEIGHT_TRUE = 52.5625 mm

Height of the A9 paper under the ISO 216 standard.

See also A9_HEIGHT_TRUE, A9_HEIGHT

MakieMaestro.Themes.A10_WIDTH — Constant

const A10_WIDTH = 26.0 mm

Width of the A10 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A10_WIDTH_TRUE, A10_HEIGHT

MakieMaestro.Themes.A10_WIDTH_TRUE — Constant

const A10_WIDTH_TRUE = 26.28125 mm

Width of the A10 paper under the ISO 216 standard.

See also A10_WIDTH_TRUE, A10_WIDTH

MakieMaestro.Themes.A10_HEIGHT — Constant

const A10_HEIGHT = 37.0 mm

Height of the A10 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A10_HEIGHT_TRUE, A10_WIDTH

MakieMaestro.Themes.A10_HEIGHT_TRUE — Constant

const A10_HEIGHT_TRUE = 37.15625 mm

Height of the A10 paper under the ISO 216 standard.

See also A10_HEIGHT_TRUE, A10_HEIGHT

MakieMaestro.Themes.A11_WIDTH — Constant

const A11_WIDTH = 18.0 mm

Width of the A11 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A11_WIDTH_TRUE, A11_HEIGHT

MakieMaestro.Themes.A11_WIDTH_TRUE — Constant

const A11_WIDTH_TRUE = 18.578125 mm

Width of the A11 paper under the ISO 216 standard.

See also A11_WIDTH_TRUE, A11_WIDTH

MakieMaestro.Themes.A11_HEIGHT — Constant

const A11_HEIGHT = 26.0 mm

Height of the A11 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A11_HEIGHT_TRUE, A11_WIDTH

MakieMaestro.Themes.A11_HEIGHT_TRUE — Constant

const A11_HEIGHT_TRUE = 26.28125 mm

Height of the A11 paper under the ISO 216 standard.

See also A11_HEIGHT_TRUE, A11_HEIGHT

MakieMaestro.Themes.A12_WIDTH — Constant

const A12_WIDTH = 13.0 mm

Width of the A12 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A12_WIDTH_TRUE, A12_HEIGHT

MakieMaestro.Themes.A12_WIDTH_TRUE — Constant

const A12_WIDTH_TRUE = 13.140625 mm

Width of the A12 paper under the ISO 216 standard.

See also A12_WIDTH_TRUE, A12_WIDTH

MakieMaestro.Themes.A12_HEIGHT — Constant

const A12_HEIGHT = 18.0 mm

Height of the A12 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A12_HEIGHT_TRUE, A12_WIDTH

MakieMaestro.Themes.A12_HEIGHT_TRUE — Constant

const A12_HEIGHT_TRUE = 18.578125 mm

Height of the A12 paper under the ISO 216 standard.

See also A12_HEIGHT_TRUE, A12_HEIGHT

MakieMaestro.Themes.A13_WIDTH — Constant

const A13_WIDTH = 9.0 mm

Width of the A13 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A13_WIDTH_TRUE, A13_HEIGHT

MakieMaestro.Themes.A13_WIDTH_TRUE — Constant

const A13_WIDTH_TRUE = 9.2890625 mm

Width of the A13 paper under the ISO 216 standard.

See also A13_WIDTH_TRUE, A13_WIDTH

MakieMaestro.Themes.A13_HEIGHT — Constant

const A13_HEIGHT = 13.0 mm

Height of the A13 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also A13_HEIGHT_TRUE, A13_WIDTH

MakieMaestro.Themes.A13_HEIGHT_TRUE — Constant

const A13_HEIGHT_TRUE = 13.140625 mm

Height of the A13 paper under the ISO 216 standard.

See also A13_HEIGHT_TRUE, A13_HEIGHT

MakieMaestro.Themes.B0_WIDTH — Constant

const B0_WIDTH = 1000.0 mm

Width of the B0 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B0_WIDTH_TRUE, B0_HEIGHT

MakieMaestro.Themes.B0_WIDTH_TRUE — Constant

const B0_WIDTH_TRUE = 1000.0 mm

Width of the B0 paper under the ISO 216 standard.

See also B0_WIDTH_TRUE, B0_WIDTH

MakieMaestro.Themes.B0_HEIGHT — Constant

const B0_HEIGHT = 1414.0 mm

Height of the B0 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B0_HEIGHT_TRUE, B0_WIDTH

MakieMaestro.Themes.B0_HEIGHT_TRUE — Constant

const B0_HEIGHT_TRUE = 1414.0 mm

Height of the B0 paper under the ISO 216 standard.

See also B0_HEIGHT_TRUE, B0_HEIGHT

MakieMaestro.Themes.B1_WIDTH — Constant

const B1_WIDTH = 707.0 mm

Width of the B1 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B1_WIDTH_TRUE, B1_HEIGHT

MakieMaestro.Themes.B1_WIDTH_TRUE — Constant

const B1_WIDTH_TRUE = 707.0 mm

Width of the B1 paper under the ISO 216 standard.

See also B1_WIDTH_TRUE, B1_WIDTH

MakieMaestro.Themes.B1_HEIGHT — Constant

const B1_HEIGHT = 1000.0 mm

Height of the B1 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B1_HEIGHT_TRUE, B1_WIDTH

MakieMaestro.Themes.B1_HEIGHT_TRUE — Constant

const B1_HEIGHT_TRUE = 1000.0 mm

Height of the B1 paper under the ISO 216 standard.

See also B1_HEIGHT_TRUE, B1_HEIGHT

MakieMaestro.Themes.B2_WIDTH — Constant

const B2_WIDTH = 500.0 mm

Width of the B2 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B2_WIDTH_TRUE, B2_HEIGHT

MakieMaestro.Themes.B2_WIDTH_TRUE — Constant

const B2_WIDTH_TRUE = 500.0 mm

Width of the B2 paper under the ISO 216 standard.

See also B2_WIDTH_TRUE, B2_WIDTH

MakieMaestro.Themes.B2_HEIGHT — Constant

const B2_HEIGHT = 707.0 mm

Height of the B2 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B2_HEIGHT_TRUE, B2_WIDTH

MakieMaestro.Themes.B2_HEIGHT_TRUE — Constant

const B2_HEIGHT_TRUE = 707.0 mm

Height of the B2 paper under the ISO 216 standard.

See also B2_HEIGHT_TRUE, B2_HEIGHT

MakieMaestro.Themes.B3_WIDTH — Constant

const B3_WIDTH = 353.0 mm

Width of the B3 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B3_WIDTH_TRUE, B3_HEIGHT

MakieMaestro.Themes.B3_WIDTH_TRUE — Constant

const B3_WIDTH_TRUE = 353.5 mm

Width of the B3 paper under the ISO 216 standard.

See also B3_WIDTH_TRUE, B3_WIDTH

MakieMaestro.Themes.B3_HEIGHT — Constant

const B3_HEIGHT = 500.0 mm

Height of the B3 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B3_HEIGHT_TRUE, B3_WIDTH

MakieMaestro.Themes.B3_HEIGHT_TRUE — Constant

const B3_HEIGHT_TRUE = 500.0 mm

Height of the B3 paper under the ISO 216 standard.

See also B3_HEIGHT_TRUE, B3_HEIGHT

MakieMaestro.Themes.B4_WIDTH — Constant

const B4_WIDTH = 250.0 mm

Width of the B4 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B4_WIDTH_TRUE, B4_HEIGHT

MakieMaestro.Themes.B4_WIDTH_TRUE — Constant

const B4_WIDTH_TRUE = 250.0 mm

Width of the B4 paper under the ISO 216 standard.

See also B4_WIDTH_TRUE, B4_WIDTH

MakieMaestro.Themes.B4_HEIGHT — Constant

const B4_HEIGHT = 353.0 mm

Height of the B4 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B4_HEIGHT_TRUE, B4_WIDTH

MakieMaestro.Themes.B4_HEIGHT_TRUE — Constant

const B4_HEIGHT_TRUE = 353.5 mm

Height of the B4 paper under the ISO 216 standard.

See also B4_HEIGHT_TRUE, B4_HEIGHT

MakieMaestro.Themes.B5_WIDTH — Constant

const B5_WIDTH = 176.0 mm

Width of the B5 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B5_WIDTH_TRUE, B5_HEIGHT

MakieMaestro.Themes.B5_WIDTH_TRUE — Constant

const B5_WIDTH_TRUE = 176.75 mm

Width of the B5 paper under the ISO 216 standard.

See also B5_WIDTH_TRUE, B5_WIDTH

MakieMaestro.Themes.B5_HEIGHT — Constant

const B5_HEIGHT = 250.0 mm

Height of the B5 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B5_HEIGHT_TRUE, B5_WIDTH

MakieMaestro.Themes.B5_HEIGHT_TRUE — Constant

const B5_HEIGHT_TRUE = 250.0 mm

Height of the B5 paper under the ISO 216 standard.

See also B5_HEIGHT_TRUE, B5_HEIGHT

MakieMaestro.Themes.B6_WIDTH — Constant

const B6_WIDTH = 125.0 mm

Width of the B6 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B6_WIDTH_TRUE, B6_HEIGHT

MakieMaestro.Themes.B6_WIDTH_TRUE — Constant

const B6_WIDTH_TRUE = 125.0 mm

Width of the B6 paper under the ISO 216 standard.

See also B6_WIDTH_TRUE, B6_WIDTH

MakieMaestro.Themes.B6_HEIGHT — Constant

const B6_HEIGHT = 176.0 mm

Height of the B6 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B6_HEIGHT_TRUE, B6_WIDTH

MakieMaestro.Themes.B6_HEIGHT_TRUE — Constant

const B6_HEIGHT_TRUE = 176.75 mm

Height of the B6 paper under the ISO 216 standard.

See also B6_HEIGHT_TRUE, B6_HEIGHT

MakieMaestro.Themes.B7_WIDTH — Constant

const B7_WIDTH = 88.0 mm

Width of the B7 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B7_WIDTH_TRUE, B7_HEIGHT

MakieMaestro.Themes.B7_WIDTH_TRUE — Constant

const B7_WIDTH_TRUE = 88.375 mm

Width of the B7 paper under the ISO 216 standard.

See also B7_WIDTH_TRUE, B7_WIDTH

MakieMaestro.Themes.B7_HEIGHT — Constant

const B7_HEIGHT = 125.0 mm

Height of the B7 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B7_HEIGHT_TRUE, B7_WIDTH

MakieMaestro.Themes.B7_HEIGHT_TRUE — Constant

const B7_HEIGHT_TRUE = 125.0 mm

Height of the B7 paper under the ISO 216 standard.

See also B7_HEIGHT_TRUE, B7_HEIGHT

MakieMaestro.Themes.B8_WIDTH — Constant

const B8_WIDTH = 62.0 mm

Width of the B8 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B8_WIDTH_TRUE, B8_HEIGHT

MakieMaestro.Themes.B8_WIDTH_TRUE — Constant

const B8_WIDTH_TRUE = 62.5 mm

Width of the B8 paper under the ISO 216 standard.

See also B8_WIDTH_TRUE, B8_WIDTH

MakieMaestro.Themes.B8_HEIGHT — Constant

const B8_HEIGHT = 88.0 mm

Height of the B8 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B8_HEIGHT_TRUE, B8_WIDTH

MakieMaestro.Themes.B8_HEIGHT_TRUE — Constant

const B8_HEIGHT_TRUE = 88.375 mm

Height of the B8 paper under the ISO 216 standard.

See also B8_HEIGHT_TRUE, B8_HEIGHT

MakieMaestro.Themes.B9_WIDTH — Constant

const B9_WIDTH = 44.0 mm

Width of the B9 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B9_WIDTH_TRUE, B9_HEIGHT

MakieMaestro.Themes.B9_WIDTH_TRUE — Constant

const B9_WIDTH_TRUE = 44.1875 mm

Width of the B9 paper under the ISO 216 standard.

See also B9_WIDTH_TRUE, B9_WIDTH

MakieMaestro.Themes.B9_HEIGHT — Constant

const B9_HEIGHT = 62.0 mm

Height of the B9 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B9_HEIGHT_TRUE, B9_WIDTH

MakieMaestro.Themes.B9_HEIGHT_TRUE — Constant

const B9_HEIGHT_TRUE = 62.5 mm

Height of the B9 paper under the ISO 216 standard.

See also B9_HEIGHT_TRUE, B9_HEIGHT

MakieMaestro.Themes.B10_WIDTH — Constant

const B10_WIDTH = 31.0 mm

Width of the B10 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B10_WIDTH_TRUE, B10_HEIGHT

MakieMaestro.Themes.B10_WIDTH_TRUE — Constant

const B10_WIDTH_TRUE = 31.25 mm

Width of the B10 paper under the ISO 216 standard.

See also B10_WIDTH_TRUE, B10_WIDTH

MakieMaestro.Themes.B10_HEIGHT — Constant

const B10_HEIGHT = 44.0 mm

Height of the B10 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B10_HEIGHT_TRUE, B10_WIDTH

MakieMaestro.Themes.B10_HEIGHT_TRUE — Constant

const B10_HEIGHT_TRUE = 44.1875 mm

Height of the B10 paper under the ISO 216 standard.

See also B10_HEIGHT_TRUE, B10_HEIGHT

MakieMaestro.Themes.B11_WIDTH — Constant

const B11_WIDTH = 22.0 mm

Width of the B11 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B11_WIDTH_TRUE, B11_HEIGHT

MakieMaestro.Themes.B11_WIDTH_TRUE — Constant

const B11_WIDTH_TRUE = 22.09375 mm

Width of the B11 paper under the ISO 216 standard.

See also B11_WIDTH_TRUE, B11_WIDTH

MakieMaestro.Themes.B11_HEIGHT — Constant

const B11_HEIGHT = 31.0 mm

Height of the B11 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B11_HEIGHT_TRUE, B11_WIDTH

MakieMaestro.Themes.B11_HEIGHT_TRUE — Constant

const B11_HEIGHT_TRUE = 31.25 mm

Height of the B11 paper under the ISO 216 standard.

See also B11_HEIGHT_TRUE, B11_HEIGHT

MakieMaestro.Themes.B12_WIDTH — Constant

const B12_WIDTH = 15.0 mm

Width of the B12 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B12_WIDTH_TRUE, B12_HEIGHT

MakieMaestro.Themes.B12_WIDTH_TRUE — Constant

const B12_WIDTH_TRUE = 15.625 mm

Width of the B12 paper under the ISO 216 standard.

See also B12_WIDTH_TRUE, B12_WIDTH

MakieMaestro.Themes.B12_HEIGHT — Constant

const B12_HEIGHT = 22.0 mm

Height of the B12 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B12_HEIGHT_TRUE, B12_WIDTH

MakieMaestro.Themes.B12_HEIGHT_TRUE — Constant

const B12_HEIGHT_TRUE = 22.09375 mm

Height of the B12 paper under the ISO 216 standard.

See also B12_HEIGHT_TRUE, B12_HEIGHT

MakieMaestro.Themes.B13_WIDTH — Constant

const B13_WIDTH = 11.0 mm

Width of the B13 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B13_WIDTH_TRUE, B13_HEIGHT

MakieMaestro.Themes.B13_WIDTH_TRUE — Constant

const B13_WIDTH_TRUE = 11.046875 mm

Width of the B13 paper under the ISO 216 standard.

See also B13_WIDTH_TRUE, B13_WIDTH

MakieMaestro.Themes.B13_HEIGHT — Constant

const B13_HEIGHT = 15.0 mm

Height of the B13 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also B13_HEIGHT_TRUE, B13_WIDTH

MakieMaestro.Themes.B13_HEIGHT_TRUE — Constant

const B13_HEIGHT_TRUE = 15.625 mm

Height of the B13 paper under the ISO 216 standard.

See also B13_HEIGHT_TRUE, B13_HEIGHT

MakieMaestro.Themes.C0_WIDTH — Constant

const C0_WIDTH = 917.0 mm

Width of the C0 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C0_WIDTH_TRUE, C0_HEIGHT

MakieMaestro.Themes.C0_WIDTH_TRUE — Constant

const C0_WIDTH_TRUE = 917.0 mm

Width of the C0 paper under the ISO 216 standard.

See also C0_WIDTH_TRUE, C0_WIDTH

MakieMaestro.Themes.C0_HEIGHT — Constant

const C0_HEIGHT = 1297.0 mm

Height of the C0 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C0_HEIGHT_TRUE, C0_WIDTH

MakieMaestro.Themes.C0_HEIGHT_TRUE — Constant

const C0_HEIGHT_TRUE = 1297.0 mm

Height of the C0 paper under the ISO 216 standard.

See also C0_HEIGHT_TRUE, C0_HEIGHT

MakieMaestro.Themes.C1_WIDTH — Constant

const C1_WIDTH = 648.0 mm

Width of the C1 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C1_WIDTH_TRUE, C1_HEIGHT

MakieMaestro.Themes.C1_WIDTH_TRUE — Constant

const C1_WIDTH_TRUE = 648.5 mm

Width of the C1 paper under the ISO 216 standard.

See also C1_WIDTH_TRUE, C1_WIDTH

MakieMaestro.Themes.C1_HEIGHT — Constant

const C1_HEIGHT = 917.0 mm

Height of the C1 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C1_HEIGHT_TRUE, C1_WIDTH

MakieMaestro.Themes.C1_HEIGHT_TRUE — Constant

const C1_HEIGHT_TRUE = 917.0 mm

Height of the C1 paper under the ISO 216 standard.

See also C1_HEIGHT_TRUE, C1_HEIGHT

MakieMaestro.Themes.C2_WIDTH — Constant

const C2_WIDTH = 458.0 mm

Width of the C2 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C2_WIDTH_TRUE, C2_HEIGHT

MakieMaestro.Themes.C2_WIDTH_TRUE — Constant

const C2_WIDTH_TRUE = 458.5 mm

Width of the C2 paper under the ISO 216 standard.

See also C2_WIDTH_TRUE, C2_WIDTH

MakieMaestro.Themes.C2_HEIGHT — Constant

const C2_HEIGHT = 648.0 mm

Height of the C2 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C2_HEIGHT_TRUE, C2_WIDTH

MakieMaestro.Themes.C2_HEIGHT_TRUE — Constant

const C2_HEIGHT_TRUE = 648.5 mm

Height of the C2 paper under the ISO 216 standard.

See also C2_HEIGHT_TRUE, C2_HEIGHT

MakieMaestro.Themes.C3_WIDTH — Constant

const C3_WIDTH = 324.0 mm

Width of the C3 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C3_WIDTH_TRUE, C3_HEIGHT

MakieMaestro.Themes.C3_WIDTH_TRUE — Constant

const C3_WIDTH_TRUE = 324.25 mm

Width of the C3 paper under the ISO 216 standard.

See also C3_WIDTH_TRUE, C3_WIDTH

MakieMaestro.Themes.C3_HEIGHT — Constant

const C3_HEIGHT = 458.0 mm

Height of the C3 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C3_HEIGHT_TRUE, C3_WIDTH

MakieMaestro.Themes.C3_HEIGHT_TRUE — Constant

const C3_HEIGHT_TRUE = 458.5 mm

Height of the C3 paper under the ISO 216 standard.

See also C3_HEIGHT_TRUE, C3_HEIGHT

MakieMaestro.Themes.C4_WIDTH — Constant

const C4_WIDTH = 229.0 mm

Width of the C4 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C4_WIDTH_TRUE, C4_HEIGHT

MakieMaestro.Themes.C4_WIDTH_TRUE — Constant

const C4_WIDTH_TRUE = 229.25 mm

Width of the C4 paper under the ISO 216 standard.

See also C4_WIDTH_TRUE, C4_WIDTH

MakieMaestro.Themes.C4_HEIGHT — Constant

const C4_HEIGHT = 324.0 mm

Height of the C4 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C4_HEIGHT_TRUE, C4_WIDTH

MakieMaestro.Themes.C4_HEIGHT_TRUE — Constant

const C4_HEIGHT_TRUE = 324.25 mm

Height of the C4 paper under the ISO 216 standard.

See also C4_HEIGHT_TRUE, C4_HEIGHT

MakieMaestro.Themes.C5_WIDTH — Constant

const C5_WIDTH = 162.0 mm

Width of the C5 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C5_WIDTH_TRUE, C5_HEIGHT

MakieMaestro.Themes.C5_WIDTH_TRUE — Constant

const C5_WIDTH_TRUE = 162.125 mm

Width of the C5 paper under the ISO 216 standard.

See also C5_WIDTH_TRUE, C5_WIDTH

MakieMaestro.Themes.C5_HEIGHT — Constant

const C5_HEIGHT = 229.0 mm

Height of the C5 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C5_HEIGHT_TRUE, C5_WIDTH

MakieMaestro.Themes.C5_HEIGHT_TRUE — Constant

const C5_HEIGHT_TRUE = 229.25 mm

Height of the C5 paper under the ISO 216 standard.

See also C5_HEIGHT_TRUE, C5_HEIGHT

MakieMaestro.Themes.C6_WIDTH — Constant

const C6_WIDTH = 114.0 mm

Width of the C6 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C6_WIDTH_TRUE, C6_HEIGHT

MakieMaestro.Themes.C6_WIDTH_TRUE — Constant

const C6_WIDTH_TRUE = 114.625 mm

Width of the C6 paper under the ISO 216 standard.

See also C6_WIDTH_TRUE, C6_WIDTH

MakieMaestro.Themes.C6_HEIGHT — Constant

const C6_HEIGHT = 162.0 mm

Height of the C6 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C6_HEIGHT_TRUE, C6_WIDTH

MakieMaestro.Themes.C6_HEIGHT_TRUE — Constant

const C6_HEIGHT_TRUE = 162.125 mm

Height of the C6 paper under the ISO 216 standard.

See also C6_HEIGHT_TRUE, C6_HEIGHT

MakieMaestro.Themes.C7_WIDTH — Constant

const C7_WIDTH = 81.0 mm

Width of the C7 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C7_WIDTH_TRUE, C7_HEIGHT

MakieMaestro.Themes.C7_WIDTH_TRUE — Constant

const C7_WIDTH_TRUE = 81.0625 mm

Width of the C7 paper under the ISO 216 standard.

See also C7_WIDTH_TRUE, C7_WIDTH

MakieMaestro.Themes.C7_HEIGHT — Constant

const C7_HEIGHT = 114.0 mm

Height of the C7 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C7_HEIGHT_TRUE, C7_WIDTH

MakieMaestro.Themes.C7_HEIGHT_TRUE — Constant

const C7_HEIGHT_TRUE = 114.625 mm

Height of the C7 paper under the ISO 216 standard.

See also C7_HEIGHT_TRUE, C7_HEIGHT

MakieMaestro.Themes.C8_WIDTH — Constant

const C8_WIDTH = 57.0 mm

Width of the C8 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C8_WIDTH_TRUE, C8_HEIGHT

MakieMaestro.Themes.C8_WIDTH_TRUE — Constant

const C8_WIDTH_TRUE = 57.3125 mm

Width of the C8 paper under the ISO 216 standard.

See also C8_WIDTH_TRUE, C8_WIDTH

MakieMaestro.Themes.C8_HEIGHT — Constant

const C8_HEIGHT = 81.0 mm

Height of the C8 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C8_HEIGHT_TRUE, C8_WIDTH

MakieMaestro.Themes.C8_HEIGHT_TRUE — Constant

const C8_HEIGHT_TRUE = 81.0625 mm

Height of the C8 paper under the ISO 216 standard.

See also C8_HEIGHT_TRUE, C8_HEIGHT

MakieMaestro.Themes.C9_WIDTH — Constant

const C9_WIDTH = 40.0 mm

Width of the C9 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C9_WIDTH_TRUE, C9_HEIGHT

MakieMaestro.Themes.C9_WIDTH_TRUE — Constant

const C9_WIDTH_TRUE = 40.53125 mm

Width of the C9 paper under the ISO 216 standard.

See also C9_WIDTH_TRUE, C9_WIDTH

MakieMaestro.Themes.C9_HEIGHT — Constant

const C9_HEIGHT = 57.0 mm

Height of the C9 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C9_HEIGHT_TRUE, C9_WIDTH

MakieMaestro.Themes.C9_HEIGHT_TRUE — Constant

const C9_HEIGHT_TRUE = 57.3125 mm

Height of the C9 paper under the ISO 216 standard.

See also C9_HEIGHT_TRUE, C9_HEIGHT

MakieMaestro.Themes.C10_WIDTH — Constant

const C10_WIDTH = 28.0 mm

Width of the C10 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C10_WIDTH_TRUE, C10_HEIGHT

MakieMaestro.Themes.C10_WIDTH_TRUE — Constant

const C10_WIDTH_TRUE = 28.65625 mm

Width of the C10 paper under the ISO 216 standard.

See also C10_WIDTH_TRUE, C10_WIDTH

MakieMaestro.Themes.C10_HEIGHT — Constant

const C10_HEIGHT = 40.0 mm

Height of the C10 paper as it is usually though of (unlike the ISO 216 standard defines it).

See also C10_HEIGHT_TRUE, C10_WIDTH

MakieMaestro.Themes.C10_HEIGHT_TRUE — Constant

const C10_HEIGHT_TRUE = 40.53125 mm

Height of the C10 paper under the ISO 216 standard.

See also C10_HEIGHT_TRUE, C10_HEIGHT

Plotting

Overrides

MakieMaestro.Recipes.image! — Function

Recipes.image!(args...; decorations=false, interpolate=false, axis=(; aspect=DataAspect()), colormap=:viridis, kwargs...)

Modified image! recipe that enables the user to set axis attributes even with the mutating function and has different defaults.

decorations – hide the axis decorations if set to false

MakieMaestro.Recipes.image — Function

Recipes.image(f)

Helper for setting the prefered options (interpolate, ax.aspect, decorations) for the image recipe.

Utility / QOL functions

MakieMaestro.L — Function

L(s::String)

Shortcut for latexstring(raw" ext{" * s * "}").

This is useful when you want to have LaTeX axis labels, but do not want to use the theme_latexfonts() theme.

MakieMaestro.fftvis — Function

fftvis(fft_img)

Apply a logarithmic scaling to visualize the absolute value of the Fourier image.

Extended help

This function takes the absolute value of the input img (typically the result of an FFT), adds 1 to avoid log(0), and then applies a base-2 logarithm. This scaling helps to visualize the wide dynamic range typically present in Fourier transforms of images.

Examples

using FFTW
img = rand(100, 100)
fft_img = fft(img)
heatmap(fftvis(fftshift(fft_img)))