Public Documentation

Documentation for MakieMaestro.jl's public interface.

Contents

Contents

Public Interface

RecipeOverrides

MakieMaestro.Themes.figsize

MakieMaestro.Themes.get_hwratio

MakieMaestro.Themes.get_theme

MakieMaestro.Themes.get_width

MakieMaestro.Themes.hwratio!

MakieMaestro.Themes.interactive_size!

MakieMaestro.Themes.merge_generate

MakieMaestro.Themes.screen_parameters

MakieMaestro.Themes.to_units

MakieMaestro.Themes.update_theme

MakieMaestro.Themes.update_theme!

MakieMaestro.Themes.width!

Public Interface

MakieMaestro.Themes.get_theme — Function

get_theme(themes::Vector{Union{ThemeGenerator,Symbol}}; dict = THEME[])
get_theme(A, B, C, ...; dict = THEME[])

Create theme for specified keys and/or generators (including Themes).

Arguments

keys::Vector{Symbol}: A vector of symbols representing the desired theme property keys.

Example

gap = true
theme_props = get_theme([Theme(; figure_padding=2), :font, () -> Theme(; colgap = gap, rowgap = gap),  :linewidth, :color])

This will return a theme that combines the three specified themes together. Note that rightmost has precedence unlike merges on usual julia dictionaries, but same as with Makie.Attributes.

MakieMaestro.Themes.interactive_size! — Function

interactive_size!(ratio::Tuple{Number,Number}; index=nothing) -> Tuple

Calculate the interactive size based on the given ratio and screen parameters.

Arguments

ratio::Tuple{Number,Number}: A tuple of two numbers between 0 and 1 representing the ratio of the screen size.

index::Union{Nothing,Int}: Optional. If provided, uses the screen parameters at the specified index.

Returns

A tuple representing the calculated interactive size.

Throws

ArgumentError: If the ratio is not between 0 and 1.

Any error that might occur when filtering for the default screen.

Description

This function calculates the interactive size by multiplying the screen size with the provided ratio. If an index is provided, it uses the screen parameters at that index. Otherwise, it attempts to use the default screen parameters.

MakieMaestro.Themes.hwratio! — Function

hwratio!(val)

Set the default height-width ratio for figures.

julia> MakieMaestro.hwratio!(0.8)

MakieMaestro.Themes.get_hwratio — Function

get_hwratio()

Get the default height-width ratio for figures

MakieMaestro.Themes.get_width — Function

get_width()

Get the default figure width

MakieMaestro.Themes.to_units — Function

to_units(val::Length)

Convert val to Makie figure units

MakieMaestro.Themes.figsize — Function

figsize(width::Length=get_width(), hw_ratio=get_hwratio())

Calculate the figure size in points based on the given width and height-to-width ratio.

Arguments

width::Length: The desired width of the figure. Defaults to the result of get_width().

hw_ratio: The height-to-width ratio. Defaults to the result of get_hwratio().

Example

width, height = figsize(800u"px", 0.75)

MakieMaestro.Themes.screen_parameters — Function

screen_parameters() -> Vector{ScreenInfo}

Retrieve information about available screens using the xdpyinfo command.

This function parses the output of xdpyinfo to extract details about each screen, including its index, dimensions in pixels and millimeters, and whether it's the default screen.

Returns:

A vector of ScreenInfo objects, each containing details about a screen.

Note:

This function relies on the xdpyinfo command and is therefore only compatible with systems where this command is available (typically Unix-like systems with X11).

The function may return an empty vector if no screens are detected or if parsing fails.

Requires the Unitful.jl package for handling millimeter units.

Example:

screens = screen_parameters()
for screen in screens
    println("Screen $(screen.index): $(screen.size_px) pixels, $(screen.size_mm) physical size")
end

MakieMaestro.Themes.update_theme — Function

update_theme(key::Symbol, with::ThemeGenerator)

Update a specific theme component identified by key in the global THEME dictionary.

Arguments

key::Symbol: The key identifying the theme component to update.

with::ThemeGenerator: The new theme or function to update the existing theme with.

See also Use update_theme! for direct modifications.

Extended help

Specific behaviour for argument types:

If the key doesn't exist in THEME, it adds the new theme or function.

If the key exists:

For an existing Theme:

If with is a Function, it merges the result of with with the current theme.

If with is a Theme, it merges the current theme with with.

For an existing Function:

If with is a Theme, it creates a new function that merges with with the result of the current function.

If with is a Function, it throws an ArgumentError.

Throws

ArgumentError: If attempting to update a generating function with another generating function.

MakieMaestro.Themes.update_theme! — Function

update_theme!(key::Symbol, new::ThemeGenerator)

Update a specific theme component in the global theme.

Arguments

key::Symbol: The key representing the theme component to be updated.

new::ThemeGenerator: The new theme generator to replace the existing one.

This function modifies the global theme by replacing the theme generator for the specified component with a new one. It directly updates the THEME global variable.

See also Use update_theme for a function a non-overwriting method.

MakieMaestro.Themes.width! — Function

width!(val)

Set the default width for figures.

julia> MakieMaestro.width!(177u"mm" * 0.8)

MakieMaestro.Themes.merge_generate — Function

merge_generate(themes::ThemeGenerator)

Merge and generate themes from a ThemeGenerator.

This function takes a ThemeGenerator (which can be a collection of themes or theme-generating functions) and merges them into a single theme. If an element of themes is a function, it is called to generate a theme; otherwise, the element is used as-is.

Arguments

themes::ThemeGenerator: A collection of themes or theme-generating functions.

Returns

A merged theme combining all the input themes. Note that the inputs that earlier in the collection have precedence.

Example

theme = merge_generate(BASE_THEME,Theme(; figure_padding=2), GL_THEME, SIZE_THEME(20u"cm", 0.5))

RecipeOverrides